simplify, deduplicate
