+++ test-vectors/CMakeLists.txt
@@ -4,4 +4,6 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
+include(asn1.cmake)
+
 add_subdirectory(safrole)
