+++ test-vectors/safrole/safrole.hpp
@@ -6,33 +6,10 @@
 
 #pragma once
 
-#include <blake2.h>
 #include <jam_crust.h>
-#include <algorithm>
-#include <boost/endian/conversion.hpp>
 
+#include "../common.hpp"
 #include "types.hpp"
-
-struct Blake {
-  using Hash = qtils::BytesN<32>;
-  blake2b_state state;
-  Blake() {
-    blake2b_init(&state, sizeof(Hash));
-  }
-  Blake &update(qtils::BytesIn input) {
-    blake2b_update(&state, input.data(), input.size());
-    return *this;
-  }
-  Hash hash() const {
-    Hash hash;
-    auto state2 = state;
-    blake2b_final(&state2, hash.data(), sizeof(Hash));
-    return hash;
-  }
-  static Hash hash(qtils::BytesIn input) {
-    return Blake{}.update(input).hash();
-  }
-};
 
 namespace jam::safrole {
   namespace generic = jam::test_vectors_safrole::generic;
@@ -40,58 +17,6 @@ namespace jam::safrole {
 
   template <typename types>
   struct Generic;
-
-  // [GP 0.3.6 3.7]
-  auto &circlearrowleft(const auto &s, auto i) {
-    return s[i % s.size()];
-  }
-
-  template <size_t N>
-  qtils::BytesN<N> first_bytes(qtils::BytesIn bytes) {
-    if (bytes.size() < N) {
-      throw std::logic_error("de");
-    }
-    qtils::BytesN<N> first;
-    memcpy(first.data(), bytes.data(), N);
-    return first;
-  }
-
-  // [GP 0.3.6 3.7.2]
-  template <size_t X, size_t Y>
-  qtils::BytesN<X + Y> frown(
-      const qtils::BytesN<X> &x, const qtils::BytesN<Y> &y) {
-    qtils::BytesN<X + Y> xy;
-    memcpy(xy.data(), x.data(), X);
-    memcpy(xy.data() + X, y.data(), Y);
-    return xy;
-  }
-
-  // [GP 0.3.6 3.7.2]
-  template <size_t X>
-  qtils::BytesN<X + 1> doubleplus(const qtils::BytesN<X> &x, uint8_t i) {
-    return frown(x, std::array{i});
-  }
-
-  // [GP 0.3.6 C.1.2 271]
-  template <size_t N>
-  qtils::BytesN<N> mathcal_E(uint64_t x) {
-    qtils::BytesN<N> out;
-    boost::endian::endian_store<uint64_t, N, boost::endian::order::little>(
-        out.data(), x);
-    return out;
-  }
-
-  // [GP 0.3.6 C.1.2 271]
-  template <size_t N>
-  auto de(qtils::BytesN<N> x) {
-    return boost::endian::
-        endian_load<uint64_t, N, boost::endian::order::little>(x.data());
-  }
-
-  // [GP 0.3.6 3.8.1]
-  inline generic::OpaqueHash mathcal_H(qtils::BytesIn m) {
-    return Blake::hash(m);
-  }
 
   // [GP 0.3.6 G 307]
   // [GP 0.3.6 G 310]
